+++ noodles-fastq/src/io/writer.rs
@@ -1,13 +1,64 @@
+//! FASTQ writer.
+
+mod builder;
 mod record;
 
 use std::io::{self, Write};
 
+pub use self::builder::Builder;
 use self::record::write_record;
 use crate::Record;
+
+pub(crate) const DEFAULT_DEFINITION_SEPARATOR: u8 = b' ';
 
 /// A FASTQ writer.
 pub struct Writer<W> {
     inner: W,
+    definition_separator: u8,
+}
+
+impl<W> Writer<W> {
+    /// Returns a reference to the underlying writer.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use std::io;
+    /// use noodles_fastq as fastq;
+    /// let writer = fastq::io::Writer::new(io::sink());
+    /// let _inner = writer.get_ref();
+    /// ```
+    pub fn get_ref(&self) -> &W {
+        &self.inner
+    }
+
+    /// Returns a mutable reference to the underlying writer.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use std::io;
+    /// use noodles_fastq as fastq;
+    /// let mut writer = fastq::io::Writer::new(io::sink());
+    /// let _inner = writer.get_mut();
+    /// ```
+    pub fn get_mut(&mut self) -> &mut W {
+        &mut self.inner
+    }
+
+    /// Unwraps and returns the underlying writer.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use std::io;
+    /// use noodles_fastq as fastq;
+    /// let writer = fastq::io::Writer::new(io::sink());
+    /// let _inner = writer.into_inner();
+    /// ```
+    pub fn into_inner(self) -> W {
+        self.inner
+    }
 }
 
 impl<W> Writer<W>
@@ -23,20 +74,10 @@ where
     /// let writer = fastq::io::Writer::new(Vec::new());
     /// ```
     pub fn new(inner: W) -> Self {
-        Self { inner }
-    }
-
-    /// Returns a reference to the underlying writer.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use noodles_fastq as fastq;
-    /// let writer = fastq::io::Writer::new(Vec::new());
-    /// assert!(writer.get_ref().is_empty());
-    /// ```
-    pub fn get_ref(&self) -> &W {
-        &self.inner
+        Self {
+            inner,
+            definition_separator: DEFAULT_DEFINITION_SEPARATOR,
+        }
     }
 
     /// Writes a FASTQ record.
@@ -56,6 +97,6 @@ where
     /// # Ok::<(), io::Error>(())
     /// ```
     pub fn write_record(&mut self, record: &Record) -> io::Result<()> {
-        write_record(&mut self.inner, record)
+        write_record(&mut self.inner, self.definition_separator, record)
     }
 }